sdk: cleanups and renames
